+++ model/jpa/src/main/java/org/keycloak/models/jpa/UserAdapter.java
@@ -147,9 +147,6 @@ public class UserAdapter implements UserModel, JpaModel<UserEntity> {
         } else if (UserModel.USERNAME.equals(name)) {
             setUsername(value);
             return;
-        } else if (UserModel.EMAIL_VERIFIED.equals(name)) {
-            setEmailVerified(Boolean.valueOf(value));
-            return;
         }
         // Remove all existing
         if (value == null) {
@@ -293,6 +290,8 @@ public class UserAdapter implements UserModel, JpaModel<UserEntity> {
             return Stream.of(user.getEmail());
         } else if (UserModel.USERNAME.equals(name)) {
             return Stream.of(user.getUsername());
+        } else if (UserModel.EMAIL_VERIFIED.equals(name)) {
+            return Stream.of(String.valueOf(user.isEmailVerified()));
         }
         return user.getAttributes().stream().filter(attribute -> Objects.equals(attribute.getName(), name)).
                 map(attribute -> attribute.getValue());