+++ services/src/main/java/org/keycloak/broker/oidc/OIDCIdentityProvider.java
@@ -25,7 +25,6 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
-import java.util.Optional;
 import java.util.stream.Stream;
 
 import jakarta.ws.rs.GET;
@@ -1128,23 +1127,6 @@ public class OIDCIdentityProvider extends AbstractOAuth2IdentityProvider<OIDCIde
             return keyStorage.reloadKeys(modelKey, new OIDCIdentityProviderPublicKeyLoader(session, getConfig()));
         }
         return false;
-    }
-
-    @Override
-    protected void setEmailVerified(UserModel user, BrokeredIdentityContext context) {
-        OIDCIdentityProviderConfig config = getConfig();
-        Map<String, Object> contextData = context.getContextData();
-        JsonWebToken token = (JsonWebToken) Optional.ofNullable(contextData.get(VALIDATED_ID_TOKEN))
-                .orElseGet(() -> contextData.get(VALIDATED_ACCESS_TOKEN));
-        Boolean emailVerified = getEmailVerifiedClaim(token);
-
-        if (Booleans.isFalse(config.isTrustEmail()) || emailVerified == null) {
-            // fallback to the default behavior if trust is disabled or there is no email_verified claim
-            super.setEmailVerified(user, context);
-            return;
-        }
-
-        user.setEmailVerified(emailVerified);
     }
 
     private Boolean getEmailVerifiedClaim(JsonWebToken token) {