+++ testsuite/integration-arquillian/tests/base/src/test/java/org/keycloak/testsuite/broker/OidcUserAttributeMapperTest.java
@@ -73,10 +73,10 @@ public class OidcUserAttributeMapperTest extends AbstractUserAttributeMapperTest
     @Test
     public void testImportEmailVerified() {
         testValueMappingForImportSyncMode(ImmutableMap.<String, List<String>>builder()
-                        .put("emailVerified", ImmutableList.<String>builder().add("false").build())
+                        .put("emailVerified", ImmutableList.<String>builder().add("true").build())
                         .build(),
                 ImmutableMap.<String, List<String>>builder()
-                        .put("emailVerified", ImmutableList.<String>builder().add("true").build())
+                        .put("emailVerified", ImmutableList.<String>builder().add("false").build())
                         .build()
         );
     }
@@ -84,10 +84,10 @@ public class OidcUserAttributeMapperTest extends AbstractUserAttributeMapperTest
     @Test
     public void testForceEmailVerified() {
         testValueMappingForForceSyncMode(ImmutableMap.<String, List<String>>builder()
-                        .put("emailVerified", ImmutableList.<String>builder().add("false").build())
+                        .put("emailVerified", ImmutableList.<String>builder().add("true").build())
                         .build(),
                 ImmutableMap.<String, List<String>>builder()
-                        .put("emailVerified", ImmutableList.<String>builder().add("true").build())
+                        .put("emailVerified", ImmutableList.<String>builder().add("false").build())
                         .build()
         );
     }